Added final part of the Portfolio Project
The main goal for the final part of this project was to get feedback from potential users and based on this make improvements, updates and new pages. We also needed to create scenarios for personas: why and how it would be convenient for different users. In addition, we created a flow map, a user flow, and a short description of what we are going to create and what the purpose of our application is.